+++ feature-settings/src/commonMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/feature/settings/di/Utils.kt
@@ -1,8 +1,5 @@
 package com.github.diegoberaldin.raccoonforlemmy.feature.settings.di
 
-import com.github.diegoberaldin.raccoonforlemmy.feature.settings.dialog.AboutDialogMviModel
 import com.github.diegoberaldin.raccoonforlemmy.feature.settings.main.SettingsMviModel
 
 expect fun getSettingsViewModel(): SettingsMviModel
-
-expect fun getAboutDialogViewModel(): AboutDialogMviModel
