+++ settings.gradle.kts
@@ -64,3 +64,4 @@ include(":unit-zoomableimage")
 include(":unit-postlist")
 include(":unit-managesubscriptions")
 include(":unit-multicommunity")
+include(":unit-about")
