+++ unit-about/src/commonMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/unit/about/di/Utils.kt
@@ -1,0 +1,5 @@
+package com.github.diegoberaldin.raccoonforlemmy.unit.about.di
+
+import com.github.diegoberaldin.raccoonforlemmy.unit.about.AboutDialogMviModel
+
+expect fun getAboutDialogViewModel(): AboutDialogMviModel
